refactor: move getting selection color to utils
